requirements.txt
aiogram==3.4.1
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.0
APScheduler==3.10.4
